gtestutils: Set coredump rlimit to 0 when calling g_test_trap_fork()
g_test_trap_fork() is deprecated, but it’s still tested in the GLib
tests. If we don’t set the coredump rlimit to 0 to prevent coredumps,
the test for it creates a coredump (which is expected). On systems with
systemd-coredump, that causes systemd-coredump to spawn and process the
coredump for a while. That slows down the tests, and results in spurious
coredump lists in the system logs.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

Fixes: #436