Define G_MSVC_SYMBOL_PREFIX correctly for ARM32
Symbols on x86 are prefixed with an underscore
but symbols on x64/ARM/ARM64 are not.

Relevant information concerning the prefixes for the architectures
can be found in the vcpkg project [1,2], where arm and arm64 builds
are part of the CI.

Specifically, _WIN64 is defined on Windows ARM64, so this issue is
only visible when building on ARM32.

[1] https://github.com/microsoft/vcpkg/blob/08e74979df124d9360343751cefcae615c2ec5ed/ports/glib/fix-arm-builds.patch
[2] https://github.com/microsoft/vcpkg/pull/6116