docs: Move ‘Notes’ section from README to NEWS
The content of the section was more appropriate to the `NEWS` file, and
no longer helpful for anyone reading the `README`. The `README` is meant
to be a quick introduction to the project and how to contribute to it,
not a set of notes for packagers and distributors. That’s what `NEWS`
is.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>