m4macros: replace obsolete macros AC_TRY_RUN and AC_TRY_LINK in glib-2.0.m4
Running autoconf 2.70 with -Wall,error on a configure.ac that uses
AM_PATH_GLIB_2_0 gives:

    configure.ac:261: warning: The macro `AC_TRY_RUN' is obsolete.
    configure.ac:261: You should run autoupdate.
    ./lib/autoconf/general.m4:2996: AC_TRY_RUN is expanded from...
    /usr/share/aclocal/glib-2.0.m4:11: AM_PATH_GLIB_2_0 is expanded from...
    configure.ac:261: the top level
    configure.ac:261: warning: The macro `AC_TRY_LINK' is obsolete.
    configure.ac:261: You should run autoupdate.
    ./lib/autoconf/general.m4:2919: AC_TRY_LINK is expanded from...
    /usr/share/aclocal/glib-2.0.m4:11: AM_PATH_GLIB_2_0 is expanded from...
    configure.ac:261: the top level

Run autoupdate on glib-2.0.m4 to change AC_TRY_RUN and AC_TRY_LINK into
the suggested alternative, and adjust the formatting a little bit.

The macros used in the alternative existed for long enough that there
shouldn't be a problem with backwards compatibility.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>