docs: Add REUSE-compliant dep5 file for the license of a few directories
This bulk-adds licensing information for a load of test directories and
the `po/` directory, since it’s impractical or impossible to add
licensing information for those files otherwise.

For the test files, in particular, adding the licensing headers would
affect the tests. Adding `.license` files alongside would cause them to
get interpreted by the test runner, which is undesirable.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1415