build: update sysprof configuration options
Sysprof recently cleaned up it's build configuration options. This is
perhaps a good time to update GLib to point at an updated commit'ish and
use the new options.

If now is not a good time to do the update, that is completely fine, but
I wanted to give you a turn-key MR nonetheless.