gobject: Avoid redundant atomics
According to the commit that introduced these
calls (4b334ef8f1393c997a2d83d), we are checking
the refcount here to avoid calling g_object_ref
when the refcount is 0, in the rare case that
notification would be triggered during finalize.

But we are now freezing notifications during
finalize, and after recent changes, we no longer call
g_object_ref for notification while a freeze is
in place.