tests: Fix TAP output from spawn-singlethread test
The output from the echo program ends up being outputted with the stdout
from the unit test in some tests. Ensure that the input to the echo
program is therefore a valid TAP comment, rather than garbage, so that
it’s safely ignored by a TAP parser looking at the output from the test.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>