Revert "gdesktopappinfo: Use `sh` rather than `gio-launch-desktop`"
A shell one-liner was enough to set GIO_LAUNCHED_DESKTOP_FILE_PID,
but ideally we also want to do the equivalent of sd_journal_stream_fd()
to set up its standard output and standard error streams.

Ideally we would call sd_journal_stream_fd() in a process that will
exec the real program, otherwise it will report the wrong process ID
in the Journal, but we can't easily do that in a forked child when
using posix_spawn() for subprocesses.

This reverts commit 2b533ca99ad07090d7090ad7389d1e85230aa618.

Signed-off-by: Simon McVittie <smcv@collabora.com>